Added an options page.
In the options page, you can set the hint characters, and the activate
key and modifier. Since yahe.js is a content script, it has to use
chrome API to access the local storage settings in the options page.

The CSS for options page is ripped from my own projects... I might
customize it later, if I feel like it.